Move the `DecryptStream` from `src/core/stream.js` and into its own file
