+++ src/core/stream.js
@@ -948,53 +948,6 @@ var PredictorStream = (function PredictorStreamClosure() {
   return PredictorStream;
 })();
 
-var DecryptStream = (function DecryptStreamClosure() {
-  // eslint-disable-next-line no-shadow
-  function DecryptStream(str, maybeLength, decrypt) {
-    this.str = str;
-    this.dict = str.dict;
-    this.decrypt = decrypt;
-    this.nextChunk = null;
-    this.initialized = false;
-
-    DecodeStream.call(this, maybeLength);
-  }
-
-  var chunkSize = 512;
-
-  DecryptStream.prototype = Object.create(DecodeStream.prototype);
-
-  DecryptStream.prototype.readBlock = function DecryptStream_readBlock() {
-    var chunk;
-    if (this.initialized) {
-      chunk = this.nextChunk;
-    } else {
-      chunk = this.str.getBytes(chunkSize);
-      this.initialized = true;
-    }
-    if (!chunk || chunk.length === 0) {
-      this.eof = true;
-      return;
-    }
-    this.nextChunk = this.str.getBytes(chunkSize);
-    var hasMoreData = this.nextChunk && this.nextChunk.length > 0;
-
-    var decrypt = this.decrypt;
-    chunk = decrypt(chunk, !hasMoreData);
-
-    var bufferLength = this.bufferLength;
-    var i,
-      n = chunk.length;
-    var buffer = this.ensureBuffer(bufferLength + n);
-    for (i = 0; i < n; i++) {
-      buffer[bufferLength++] = chunk[i];
-    }
-    this.bufferLength = bufferLength;
-  };
-
-  return DecryptStream;
-})();
-
 var NullStream = (function NullStreamClosure() {
   // eslint-disable-next-line no-shadow
   function NullStream() {
@@ -1008,7 +961,6 @@ var NullStream = (function NullStreamClosure() {
 
 export {
   DecodeStream,
-  DecryptStream,
   FlateStream,
   NullStream,
   PredictorStream,
